ns for fx: make weights comparison work on N models
Summary:

Refactor the NS for FX compare weights API to be able to
work on N models and do arbitrary matching strategies.

We factor out a util which takes a model and a list of
nodes to extract weights for, with names to give the extracted
weights.  The user can then call this util with a set
of nodes and names created in any way they want.

Test Plan:

```
python test/test_quantization.py TestFXNumericSuiteCoreAPIs
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 846856af35770dfe83a65306368c1ac5e1bf41b9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52356